git-svn-id: https://plugins.svn.wordpress.org/really-simple-facebook-twitter-share-buttons/trunk@351828 b8457f37-d9ea-0310-8a92-e5e31aec5664

+++ uninstall.php
@@ -1,22 +1,4 @@
-<?php
-/*
-Plugin Name: Really simple Facebook Twitter share buttons
-Plugin URI: http://www.whiletrue.it
-Description: Puts Facebook and Twitter share buttons above or below your posts.
-Author: WhileTrue
-Version: 1.0.0
-Author URI: http://www.whiletrue.it
-*/
-
-/*
-    This program is free software; you can redistribute it and/or modify
-    it under the terms of the GNU General Public License version 2, 
-    as published by the Free Software Foundation.
-
-    This program is distributed in the hope that it will be useful,
-    but WITHOUT ANY WARRANTY; without even the implied warranty of
-    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-    GNU General Public License for more details.
-*/
+if( !defined( 'ABSPATH') && !defined('WP_UNINSTALL_PLUGIN') )
+    exit();
 
 delete_option('really_simple_share');
